# Optimistically check the mode of the sensor before setting it on read