--- a/tests/api_tests.py
+++ b/tests/api_tests.py
@@ -15,6 +15,17 @@ from ev3dev2.motor import Motor, MediumMotor, MoveTank, MoveSteering, MoveJoysti
 
 ev3dev2.Device.DEVICE_ROOT_PATH = os.path.join(FAKE_SYS, 'arena')
 
+_internal_set_attribute = ev3dev2.Device._set_attribute
+def _set_attribute(self, attribute, name, value):
+    # After writing, clear the rest of the file to remove any residual text from
+    # the last write. On the real device we're writing to sysfs attributes where
+    # there isn't any persistent buffer, but in the test environment they're
+    # normal files on disk.
+    attribute = _internal_set_attribute(self, attribute, name, value)
+    attribute.truncate()
+
+ev3dev2.Device._set_attribute = _set_attribute
+
 def dummy_wait(self, cond, timeout=None):
     pass
 
@@ -100,11 +111,11 @@ class TestAPI(unittest.TestCase):
 
         val = s.proximity
         # Our test environment writes to actual files on disk, so while "seek(0) write(...)" works on the real device, it leaves trailing characters from previous writes in tests. "s.mode" returns "IR-PROXTE" here.
-        self.assertTrue(s.mode.startswith("IR-PROX"))
+        self.assertEqual(s.mode, "IR-PROX")
         self.assertEqual(val,               16)
 
         val = s.buttons_pressed()
-        self.assertTrue(s.mode.startswith("IR-REMOTE"))
+        self.assertEqual(s.mode, "IR-REMOTE")
         self.assertEqual(val,               [])
 
     def test_medium_motor_write(self):