As we've [recently discovered](https://github.com/ev3dev/ev3dev-lang-python/issues/498), some sensors don't respond well to having their mode set repeatedly<sup>1</sup>. Here's my proposed fix, which I think simultaneously improves sensor operation all-around.

<sup>1: _My guess is that this was intentional on LEGO's part; the beacon button in particular has a delay on it, presumably added on the sensor side to be resilient to momentary losses of line-of-sight. That timer is probably reset every time we write the mode._</sup>

---

I did some benchmarking with an IR sensor (UART):

```
robot@ev3dev:~/ev3dev-lang-python$ python3 -m timeit -s "import ev3dev2.sensor; s = ev3dev2.sensor.Sensor(ev3dev2.sensor.INPUT_3)" "s.mode"
1000 loops, best of 3: 975 usec per loop

robot@ev3dev:~/ev3dev-lang-python$ python3 -m timeit -s "import ev3dev2.sensor; s = ev3dev2.sensor.Sensor(ev3dev2.sensor.INPUT_3)" "s.mode = 'IR-REMOTE'"
10 loops, best of 3: 97.8 msec per loop
```

In other words, reading the mode takes around 1ms, and writing takes 100x that. This is generally what I would expect, given that setting the mode on a UART sensor always requires writing to the serial line and might even wait for a response... I haven't looked into it past a cursory glance at the top-level "set" method. I believe reading is just a memory read (+ sysfs overhead).

For comparison, here's an analog sensor (the only one IIRC), which only has one mode and presumably no-ops when setting the mode anyway (I didn't bother looking up the driver code for this one):

```
robot@ev3dev:~/ev3dev-lang-python$ python3 -m timeit -s "import ev3dev2.sensor; s = ev3dev2.sensor.Sensor(ev3dev2.sensor.INPUT_3)" "s.mode"
1000 loops, best of 3: 834 usec per loop

robot@ev3dev:~/ev3dev-lang-python$ python3 -m timeit -s "import ev3dev2.sensor; s = ev3dev2.sensor.Sensor(ev3dev2.sensor.INPUT_3)" "s.mode = 'TOUCH'"
1000 loops, best of 3: 827 usec per loop
````

So, this one is around 1ms for both operations. I'm going to guess that this is just sysfs file I/O overhead.

I didn't see much use in finding an I2C sensor, given that pretty much all of the EV3 ones are UART. It's probably somewhere in between.

---

What this tells me is that we wait 100ms every time you read a sensor, for no good reason. Given that reading the mode is only 1ms, it seems reasonable to me that we should do the read and only write if we need it to change. In most cases, that's probably the right assumption. Technically, this does introduce a potential race condition, but I think the risk is minor.

 I've included that in this PR, and added some related tests.

Fixes #498